+++ Layout/Waher.Layout.Layout2D/Model/Figures/FigurePoint.cs
@@ -79,12 +79,12 @@ namespace Waher.Layout.Layout2D.Model.Figures
 		/// <summary>
 		/// Measured X-coordinate
 		/// </summary>
-		protected double xCoordinate;
+		protected float xCoordinate;
 
 		/// <summary>
 		/// Measured Y-coordinate
 		/// </summary>
-		protected double yCoordinate;
+		protected float yCoordinate;
 
 	}
 }
