+++ Layout/Waher.Layout.Layout2D/Model/Figures/FigurePoint2.cs
@@ -80,12 +80,12 @@ namespace Waher.Layout.Layout2D.Model.Figures
 		/// <summary>
 		/// Measured X-coordinate
 		/// </summary>
-		protected double xCoordinate2;
+		protected float xCoordinate2;
 
 		/// <summary>
 		/// Measured Y-coordinate
 		/// </summary>
-		protected double yCoordinate2;
+		protected float yCoordinate2;
 
 	}
 }
